tests/wpt/web-platform-tests/css/css-multicol/parsing/column-gap-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Multi-column Layout: parsing column-gap with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-gap">
<meta name="assert" content="column-gap supports only the grammar '<length-percentage> | normal'.">
<meta name="assert" content="column-gap rejects negative <length-percentage>.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("column-gap", "auto");

test_invalid_value("column-gap", "10");
test_invalid_value("column-gap", "10px 20px");
test_invalid_value("column-gap", "-1px");
test_invalid_value("column-gap", "-10%");
</script>
</body>
</html>
